\name{RdeGlmLassoGam-package}
\alias{RdeGlmLassoGam-package}
\docType{package}
\title{
Robust Double Exponential Estimator
}
\description{
This package provides a robust estimator for the parameters in the double exponential family, the so-called Robust Double Exponential (RDE) Estimator. One can use (a combination of) Generalized Linear Models, Generalized Penalized Linear Models or Generalized Additive Models.
}
\details{
\tabular{ll}{
Package: \tab RdeGlmLassoGam\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
License: \tab GPL (>= 2)\cr
}
}


